+++ modules/portal/portal-search-lucene/src/com/liferay/portal/search/lucene/internal/LuceneFields.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.portal.search.lucene;
+package com.liferay.portal.search.lucene.internal;
 
 import com.liferay.portal.kernel.util.GetterUtil;
