+++ modules/portal/portal-search-lucene/src/com/liferay/portal/search/lucene/internal/LuceneHelperImpl.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.portal.search.lucene;
+package com.liferay.portal.search.lucene.internal;
 
 import com.liferay.portal.kernel.concurrent.ThreadPoolExecutor;
 import com.liferay.portal.kernel.executor.PortalExecutorManagerUtil;
@@ -31,7 +31,7 @@ import com.liferay.portal.kernel.util.StringPool;
 import com.liferay.portal.kernel.util.StringUtil;
 import com.liferay.portal.kernel.util.Validator;
 import com.liferay.portal.search.SearchEngineInitializer;
-import com.liferay.portal.search.lucene.highlight.QueryTermExtractor;
+import com.liferay.portal.search.lucene.internal.highlight.QueryTermExtractor;
 import com.liferay.portal.util.PropsUtil;
 import com.liferay.portal.util.PropsValues;
