+++ modules/portal/portal-search-lucene/src/com/liferay/portal/search/lucene/internal/LuceneIndexSearcher.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.portal.search.lucene;
+package com.liferay.portal.search.lucene.internal;
 
 import com.browseengine.bobo.api.BoboBrowser;
 import com.browseengine.bobo.api.BoboIndexReader;
@@ -61,7 +61,7 @@ import com.liferay.portal.kernel.util.StringPool;
 import com.liferay.portal.kernel.util.StringUtil;
 import com.liferay.portal.kernel.util.Time;
 import com.liferay.portal.kernel.util.Validator;
-import com.liferay.portal.search.BoboFacetCollector;
+import com.liferay.portal.search.lucene.QueryTranslator;
 import com.liferay.portal.util.PropsValues;
 
 import java.io.IOException;
