+++ modules/portal/portal-search-lucene/src/com/liferay/portal/search/lucene/internal/LuceneSearchEngine.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.portal.search.lucene;
+package com.liferay.portal.search.lucene.internal;
 
 import com.liferay.portal.kernel.search.BaseSearchEngine;
 import com.liferay.portal.kernel.util.FileUtil;
