+++ modules/portal/portal-search-lucene/src/com/liferay/portal/search/lucene/internal/StringDistanceRelevancyChecker.java
@@ -12,18 +12,44 @@
  * details.
  */
 
-package com.liferay.portal.search.lucene;
+package com.liferay.portal.search.lucene.internal;
 
+import org.apache.lucene.search.spell.StringDistance;
 import org.apache.lucene.search.spell.SuggestWord;
 
 /**
  * @author Michael C. Han
  */
-public class DefaultRelevancyChecker implements RelevancyChecker {
+public class StringDistanceRelevancyChecker implements RelevancyChecker {
+
+	public StringDistanceRelevancyChecker(
+		String word, float scoresThreshold, StringDistance stringDistance) {
+
+		_word = word;
+		_scoresThreshold = scoresThreshold;
+		_stringDistance = stringDistance;
+	}
 
 	@Override
 	public boolean isRelevant(SuggestWord suggestWord) {
+		String word = suggestWord.string;
+
+		if (word.equals(_word)) {
+			return false;
+		}
+
+		suggestWord.score = _stringDistance.getDistance(
+			_word, suggestWord.string);
+
+		if (suggestWord.score <= _scoresThreshold) {
+			return false;
+		}
+
 		return true;
 	}
+
+	private final float _scoresThreshold;
+	private final StringDistance _stringDistance;
+	private final String _word;
 
 }
