+++ modules/portal/portal-search-lucene/src/com/liferay/portal/search/lucene/internal/dump/DumpIndexDeletionPolicy.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.portal.search.lucene.dump;
+package com.liferay.portal.search.lucene.internal.dump;
 
 import java.io.IOException;
 import java.io.OutputStream;
