+++ modules/portal/portal-search-lucene/src/com/liferay/portal/search/lucene/internal/query/QueryTranslatorImpl.java
@@ -18,8 +18,8 @@ import com.liferay.portal.kernel.search.ParseException;
 import com.liferay.portal.kernel.search.Query;
 import com.liferay.portal.kernel.search.StringQueryImpl;
 import com.liferay.portal.kernel.util.StringPool;
-import com.liferay.portal.search.lucene.LuceneHelper;
 import com.liferay.portal.search.lucene.QueryTranslator;
+import com.liferay.portal.search.lucene.internal.LuceneHelper;
 
 import org.apache.lucene.queryParser.QueryParser;
